Update ConnectionManager.java
Changed Nodes list from LinkedList to Queue to automate cycling of Nodes.  Need to change Node.compareTo to use not only ping but also usage.  Could be based on LocalTime, or could possibly be based on int iterator, for example usageCount <1-3>, 1 is prioritized, and incremented each use until it hits 3, which is lowest priority that then cycles back to 1?